change package names to lowercase
