+++ app/src/androidTest/java/io/mochahub/powermeter/RoomRelationTest.kt
@@ -4,14 +4,14 @@ import android.content.Context
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import io.mochahub.powermeter.data.AppDatabase
-import io.mochahub.powermeter.data.Exercise.ExerciseDao
-import io.mochahub.powermeter.data.Exercise.ExerciseEntity
-import io.mochahub.powermeter.data.Workout.WorkoutDao
-import io.mochahub.powermeter.data.Workout.WorkoutEntity
-import io.mochahub.powermeter.data.WorkoutSession.WorkoutSessionDao
-import io.mochahub.powermeter.data.WorkoutSession.WorkoutSessionEntity
-import io.mochahub.powermeter.data.WorkoutSet.WorkoutSetDao
-import io.mochahub.powermeter.data.WorkoutSet.WorkoutSetEntity
+import io.mochahub.powermeter.data.exercise.ExerciseDao
+import io.mochahub.powermeter.data.exercise.ExerciseEntity
+import io.mochahub.powermeter.data.workout.WorkoutDao
+import io.mochahub.powermeter.data.workout.WorkoutEntity
+import io.mochahub.powermeter.data.workoutsession.WorkoutSessionDao
+import io.mochahub.powermeter.data.workoutsession.WorkoutSessionEntity
+import io.mochahub.powermeter.data.workoutset.WorkoutSetDao
+import io.mochahub.powermeter.data.workoutset.WorkoutSetEntity
 import kotlinx.coroutines.flow.take
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.runBlocking
