+++ app/src/main/java/io/mochahub/powermeter/data/exercise/ExerciseDao.kt
@@ -1,4 +1,4 @@
-package io.mochahub.powermeter.data.Exercise
+package io.mochahub.powermeter.data.exercise
 
 import androidx.room.Dao
 import androidx.room.Delete
