+++ app/src/main/java/io/mochahub/powermeter/workoutsession/WorkoutSessionController.kt
@@ -1,7 +1,7 @@
 package io.mochahub.powermeter.workoutsession
 
 import com.airbnb.epoxy.TypedEpoxyController
-import io.mochahub.powermeter.data.WorkoutSession.WorkoutSessionEntity
+import io.mochahub.powermeter.data.workoutsession.WorkoutSessionEntity
 
 class WorkoutSessionController(
     private val clickListener: (WorkoutSessionEntity) -> Unit
